Add basic pipeline structures

+++ 3D-Graphics/Vec.cs
@@ -4,20 +4,20 @@ using MathNet.Numerics.LinearAlgebra;
 namespace _3D_Graphics {
     public struct Vec2 {
         static Random Rnd = new Random();
-        public double X, Y;
+        public float X, Y;
 
-        public Vec2(Vector<double> mathV) {
+        public Vec2(Vector<float> mathV) {
             X = mathV[0];
             Y = mathV[1];
         }
 
-        public Vec2(double x, double y) {
+        public Vec2(float x, float y) {
             X = x;
             Y = y;
         }
 
-        public double Length() {
-            return Math.Sqrt(X * X + Y * Y);
+        public float Length() {
+            return (float)Math.Sqrt(X * X + Y * Y);
         }
 
         public Vec2 Normalize() {
@@ -25,7 +25,7 @@ namespace _3D_Graphics {
         }
 
         public static Vec2 RandomNormal() {
-            Vec2 u = new Vec2(Rnd.NextDouble() * 2.0 - 1.0, Rnd.NextDouble() * 2.0 - 1.0);
+            Vec2 u = new Vec2((float)Rnd.NextDouble() * 2.0f - 1.0f, (float)Rnd.NextDouble() * 2.0f - 1.0f);
             return u.Normalize();
         }
 
@@ -33,31 +33,31 @@ namespace _3D_Graphics {
         public static Vec2 operator -(Vec2 u) => new Vec2(-u.X, -u.Y);
         public static Vec2 operator +(Vec2 u, Vec2 v) => new Vec2(u.X + v.X, u.Y + v.Y);
         public static Vec2 operator -(Vec2 u, Vec2 v) => new Vec2(u.X - v.X, u.Y - v.Y);
-        public static Vec2 operator *(Vec2 u, double a) => new Vec2(a * u.X, a * u.Y);
-        public static Vec2 operator *(double a, Vec2 u) => u * a;
-        public static Vec2 operator /(Vec2 u, double a) => new Vec2(u.X / a, u.Y / a);
+        public static Vec2 operator *(Vec2 u, float a) => new Vec2(a * u.X, a * u.Y);
+        public static Vec2 operator *(float a, Vec2 u) => u * a;
+        public static Vec2 operator /(Vec2 u, float a) => new Vec2(u.X / a, u.Y / a);
     }
 
     public struct Vec3 {
         static Random Rnd = new Random();
-        public double X, Y, Z;
+        public float X, Y, Z;
 
-        public Vec3(double x, double y, double z) {
+        public Vec3(float x, float y, float z) {
             X = x;
             Y = y;
             Z = z;
         }
 
-        public Vec3(Vector<double> mathV) {
+        public Vec3(Vector<float> mathV) {
             X = mathV[0];
             Y = mathV[1];
             Z = mathV[2];
         }
 
         public Vec3(UInt32 color) {
-            X = (double)((color & 0x00FF0000) >> 16) / 255.0;
-            Y = (double)((color & 0x0000FF00) >> 8) / 255.0;
-            Z = (double)(color & 0x000000FF) / 255.0;
+            X = (float)((color & 0x00FF0000) >> 16) / 255.0f;
+            Y = (float)((color & 0x0000FF00) >> 8) / 255.0f;
+            Z = (float)(color & 0x000000FF) / 255.0f;
         }
 
         public UInt32 ToColor() {
@@ -69,8 +69,8 @@ namespace _3D_Graphics {
             return a | r | g | b;
         }
 
-        public double Length() {
-            return Math.Sqrt(X * X + Y * Y + Z * Z);
+        public float Length() {
+            return (float)Math.Sqrt(X * X + Y * Y + Z * Z);
         }
 
         public Vec3 Normalize() {
@@ -79,18 +79,18 @@ namespace _3D_Graphics {
 
         public static Vec3 RandomNormal() {
             Vec3 u = new Vec3(
-                Rnd.NextDouble() * 2.0 - 1.0,
-                Rnd.NextDouble() * 2.0 - 1.0,
-                Rnd.NextDouble() * 2.0 - 1.0
+                (float)Rnd.NextDouble() * 2.0f - 1.0f,
+                (float)Rnd.NextDouble() * 2.0f - 1.0f,
+                (float)Rnd.NextDouble() * 2.0f - 1.0f
             );
             return u.Normalize();
         }
 
-        public static double AngleCosine(Vec3 u, Vec3 v) {
+        public static float AngleCosine(Vec3 u, Vec3 v) {
             return DotProduct(u, v) / (u.Length() * v.Length());
         }
 
-        public static double DotProduct(Vec3 u, Vec3 v) {
+        public static float DotProduct(Vec3 u, Vec3 v) {
             return u.X * v.X + u.Y * v.Y + u.Z * v.Z;
         }
 
@@ -102,8 +102,8 @@ namespace _3D_Graphics {
         public static Vec3 operator -(Vec3 u) => new Vec3(-u.X, -u.Y, -u.Z);
         public static Vec3 operator +(Vec3 u, Vec3 v) => new Vec3(u.X + v.X, u.Y + v.Y, u.Z + v.Z);
         public static Vec3 operator -(Vec3 u, Vec3 v) => new Vec3(u.X - v.X, u.Y - v.Y, u.Z - v.Z);
-        public static Vec3 operator *(Vec3 u, double a) => new Vec3(a * u.X, a * u.Y, a * u.Z);
-        public static Vec3 operator *(double a, Vec3 u) => u * a;
-        public static Vec3 operator /(Vec3 u, double a) => new Vec3(u.X / a, u.Y / a, u.Z / a);
+        public static Vec3 operator *(Vec3 u, float a) => new Vec3(a * u.X, a * u.Y, a * u.Z);
+        public static Vec3 operator *(float a, Vec3 u) => u * a;
+        public static Vec3 operator /(Vec3 u, float a) => new Vec3(u.X / a, u.Y / a, u.Z / a);
     }
 }
